util: systemp: set pipefail flag
This ensures that the whole call to `systemp` fails if one commands in a
command line consisting of multiple commands joined by pipes fails.

This is for example relevant for the `cpio` backend, where the `cpio`
command is called and its output is piped to `gzip`.
If the `cpio` command fails or is not found at all a `.cpio.gz` file
will be created containing just the gzip header and genimage will
return a successful exit status anyways.

This change assumes that all relevant `/bin/sh` implementations support
the `-o pipefail` option. We know that at least bash, dash and busybox ash
do.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>